# server/openapi.yaml
openapi: 3.1.0
info:
  title: Reprise app API
  version: 0.1.0
  description: |
    Backend API for the Reprise Review UI (issue #30).
    Auth: POST /api/session with a GitHub token sets the httpOnly `reprise_sid` cookie.
    Long-running operations return 202 { runId }; follow GET /api/runs/{id}/events (SSE)
    or poll GET /api/runs/{id}. TypeScript types: server/src/api/types.ts.
servers:
  - url: /
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: reprise_sid
  parameters:
...
    runId: { name: id, in: path, required: true, schema: { type: string } }
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
  schemas:
    ApiError:
      type: object
      required: [error]
      properties: { error: { type: string } }
...
    IssueRecord:
      description: Schema-3 issue record (core/src/contracts/records.ts, schemas/issue-record.schema.json).
      type: object
      required: [schema, repo, issue, title, url, state, replication, fix, events]
      properties:
        schema: { const: 3 }
        repo: { type: string }
        issue: { type: integer }
        title: { type: string }
        url: { type: string }
        state: { type: string }
        replication: { type: object }
        fix: { type: object }
        events: { type: array }
    Location:
      type: object
      required: [file, start_line, end_line, reason]
      properties:
        file: { type: string }
        start_line: { type: integer }
        end_line: { type: integer }
        reason: { type: string }
    Proposal:
      type: object
      required: [locations, root_cause, fix_direction, confidence, diff, pr_draft]
      properties:
        locations: { type: array, items: { $ref: '#/components/schemas/Location' } }
        root_cause: { type: string }
        fix_direction: { type: string }
        confidence: { enum: [high, medium, low] }
        diff: { type: string, description: Unified diff }
        pr_draft:
          type: object
          required: [title, body]
          properties: { title: { type: string }, body: { type: string } }
...
    LocalCheck:
      description: The fix applied on a runner worktree (overlay at base_sha) and tested there.
      type: object
      required: [base_sha, overlay_id, files, repro, regression, verdict]
      properties:
        base_sha: { type: string }
        overlay_id: { type: string }
        files:
          type: array
          items: { type: object, required: [path, sha256], properties: { path: { type: string }, sha256: { type: string } } }
        repro:
          type: object
          required: [test_file, runs, failed, fixed]
          properties:
            test_file: { type: string }
            runs: { type: integer }
            failed: { type: integer }
            fixed: { type: boolean }
        regression:
          oneOf:
            - type: 'null'
            - type: object
              required: [tests_total, counts, blocking, notable]
              properties:
                tests_total: { type: integer }
                counts: { type: object }
                blocking: { type: array, items: { type: string } }
                notable: { type: array, items: { type: string } }
        verdict: { enum: [FIX_VERIFIED, FIX_INCOMPLETE, REGRESSION_DETECTED] }
    RunnerCall:
      description: A call the browser makes to its paired runner for the backend. Browsers must refuse anything else.
      oneOf:
        - type: object
          required: [method, path]
          properties: { method: { const: GET }, path: { type: string, pattern: '^/(status$|file\?)' } }
        - type: object
          required: [method, path, body]
          properties:
            method: { const: POST }
            path: { enum: [/approve, /overlays] }
            body: { type: object }
...
    RunStatus:
      type: object
      required: [id, kind, state, repo, issue, started_at, finished_at, record, proposal, check, error]
      properties:
        id: { type: string }
        kind: { enum: [acknowledge, propose, check] }
        state: { enum: [running, succeeded, failed] }
        repo: { type: string }
        issue: { type: integer }
        started_at: { type: string }
        finished_at: { type: [string, 'null'] }
        record: { oneOf: [{ $ref: '#/components/schemas/IssueRecord' }, { type: 'null' }] }
        proposal: { oneOf: [{ $ref: '#/components/schemas/Proposal' }, { type: 'null' }] }
        check: { oneOf: [{ $ref: '#/components/schemas/LocalCheck' }, { type: 'null' }] }
        error: { type: [string, 'null'] }
    CoreEvent:
      description: CoreEvent from core/src/contracts/services.ts.
      type: object
      required: [type]
      properties:
        type: { enum: [record.updated, status, info, error] }
    RunRequest:
      description: RunRequest from core/src/contracts/execution.ts. Forward it verbatim to the runner's POST /runs.
      type: object
      required: [platform, mode, test_path, runs, ref]
    RunStreamEvent:
      oneOf:
        - type: object
          required: [type, event]
          properties: { type: { const: core }, event: { $ref: '#/components/schemas/CoreEvent' } }
        - type: object
          required: [type, reqId, request]
          properties: { type: { const: exec.request }, reqId: { type: string }, request: { $ref: '#/components/schemas/RunRequest' } }
        - type: object
          required: [type, reqId, call]
          properties: { type: { const: runner.request }, reqId: { type: string }, call: { $ref: '#/components/schemas/RunnerCall' } }
        - type: object
          required: [type, reqId, line]
          properties: { type: { const: exec.output }, reqId: { type: string }, line: { type: string } }
        - type: object
          required: [type, status]
          properties: { type: { enum: [run.done, run.failed] }, status: { $ref: '#/components/schemas/RunStatus' } }
    ExecResult:
      oneOf:
        - type: object
          description: Answers an exec.request.
          required: [ok, results]
          properties: { ok: { const: true }, results: { type: array, items: { type: object } } }
        - type: object
          description: Answers a runner.request with the runner's HTTP status and JSON body.
          required: [ok, status, body]
          properties: { ok: { const: true }, status: { type: integer }, body: {} }
        - type: object
          required: [ok, error]
          properties: { ok: { const: false }, error: { type: string } }
...
security:
  - session: []
paths:
...
  /api/runs/{id}:
    parameters: [{ $ref: '#/components/parameters/runId' }]
    get:
      responses:
        '200': { description: Run status, content: { application/json: { schema: { $ref: '#/components/schemas/RunStatus' } } } }
        '404': { $ref: '#/components/responses/Error' }
  /api/runs/{id}/events:
    parameters: [{ $ref: '#/components/parameters/runId' }]
    get:
      description: Server-sent events. Replays buffered events, then streams live ones. Closes after run.done or run.failed.
      responses:
        '200': { description: 'text/event-stream; each data line is a RunStreamEvent', content: { text/event-stream: { schema: { $ref: '#/components/schemas/RunStreamEvent' } } } }
        '404': { $ref: '#/components/responses/Error' }
  /api/runs/{id}/exec/{reqId}:
    parameters: [{ $ref: '#/components/parameters/runId' }, { name: reqId, in: path, required: true, schema: { type: string } }]
    post:
      description: The browser posts the local runner's answer to an exec.request or runner.request.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ExecResult' } } }
      responses:
        '204': { description: Accepted }
        '400': { description: The answer does not match the request's kind, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '404': { $ref: '#/components/responses/Error' }